+++ CHANGELOG.md
@@ -19,9 +19,93 @@ __________ DO NOT TOUCH ___________
 
 __________ DO NOT TOUCH ___________
 
-## [22.3.0]
+## [22.9.0]
 ### Added
 - Samples that lack data during delivery is reported
+
+
+## [22.8.6]
+### Fixed
+- Statina load use latest file
+
+## [22.8.5]
+### Fixed
+- Sftp upload before sending email to FOHM
+
+## [22.8.4]
+### Fixed
+- Tooltips for fohm upload functions
+
+
+## [22.8.3]
+### Added
+- Data delivery export to LIMS 
+
+## [22.8.2]
+### Fixed
+- Fail graceful if Avatar url lookup fails 
+
+## [22.8.1]
+### Changed
+- Upload gisaid uses once again single fasta
+- Upload gisaid uses pandas
+- Upload gisaid uploads only qc pass files
+
+## [22.8.0]
+### Added
+- Added support for FOHM batch upload
+
+
+
+## [22.7.0]
+### Added
+- email functionality
+
+### Changed
+- gisaid upload command: gisaid cli log file will be saved in housekeeper with tag gisaid-log
+- gisaid upload command: the log file will be appended to if the upload is run again for the same case.
+- gisaid upload command: Accession numbers will be parsed from the log file 
+- gisaid upload command: The completion file which already exists in housekeeper will be updated with accession numbers
+- gisaid upload command: If files are missing or not all samples within the case were uploaded to gisiad, an email will be sent to logwatch.
+
+## [22.6.1]
+### Changed
+- Add deepvariant for input file in upload to loqusDB
+
+## [22.6.0]
+### Added
+- Added rsync processes into trailblazer
+- Added cleanup of rsync directories
+
+## [22.5.2]
+### Fixed
+- mip check if analysis is complete
+
+
+## [22.5.1]
+### Fixed
+- Avatar name lookup problem 
+
+## [22.5.0]
+### Added
+- Avatars for cases
+
+## [22.4.1]
+### Fixed
+- Statina upload api call to raise error when server response is not ok
+
+## [22.4.0]
+### Fixed
+- cg will no longer run rsync on the gpu nodes
+
+## [22.3.1]
+### Fixed
+- Fixed statina load bug where missing headers prevented data from being read
+
+## [22.3.0]
+### Added
+- Add a command to upload both to Statina and ftp 
+- Adds a command to run the first command for all available cases
 
 ## [22.2.2]
 ### Changed
@@ -99,7 +183,7 @@ __________ DO NOT TOUCH ___________
 ### Fixed
 - gisaid get region and lab from sample udfs
 - gisaid adjusting for new fasta consensus file in housekeeper
-- gisaid removing failing tests 
+- gisaid removing failing tests
 
 ### Changed
 ### Fixed
@@ -138,7 +222,6 @@ __________ DO NOT TOUCH ___________
 ## [21.7.1]
 ### Fixed
 - Fixed bug when checking if flowcell projects has been renamed
-
 
 ## [21.7.0]
 ### Added
@@ -261,14 +344,12 @@ Add uploaded to vogue date to analysis table
 Only select potential analyses to upload that have not been uploaded
 
 ## [21.0.0]
-
 ### Changed
 - Add support for balsamic 7.x.x
 - Rework Balsamic server configurations
 
 ### Fixed
 - Upload to scout now possible for all analysis types through cg upload scout
-
 
 ## [20.26.2]
 ### Added
@@ -296,6 +377,7 @@ Only select potential analyses to upload that have not been uploaded
 - gisaid uppload support via cli
 - gisaid API
 
+## [20.22.0]
 ### Added
 - Add command `cg get analysis` to view analysis information